Fixing incorrect ambiguous check.
It seems that bytecode allows for multiple occurrences of the same
placeholder and it's valid. We still want to protect ourselves for cases
where ambiguity occurs. This patch handles multiple occurrences of
placeholders which end with `___`, ie. they look like non ambiguous.

This is not ideal because it won't handle correctly names that happen
to have `_` character in the filename at 35th character.

This can be fixed further to handle this case, we could extract those
checks to helpers like `isSafePlaceholder(...)` which checks if 35th
char is different than `_` and allow replace-all on those links and
throw `Can't decide if placeholder is ambiguous.` if it's not and there
are multiple occurrences.